Add support for being lazy when using otw.
If download is set we will not be lazy.

But otherwise, only add those tiddlers to the space which are
structural, not content.

Note that this leaves you with a tiddlywiki which will not load
the tiddlers that are missing. To get that working, LoadMissingTiddlersPlugin
or something similar needs to be baked in, which is probably the
next step (can add that to otw.js).

It will likely need to be made a bit more generic?

Not sure.